pcb layout
This proposed system is very effective in the field for irrigation, Environmental parameters monitoring and predator detection. The prediction of the crop in the field is also very accurate due to the use of ph sensor. Due, to the prediction of the crop to be grown the productivity of the crop will be more. This is very useful for the farmers for monitoring the situation of the field without going to the field. This system is very much helpful for the rice crops as it requires much water than the other crops. This system also will be useful for any type of crop. The developed system is of low cost and consumes less power due to the usage of NodeMCU as micro controller.The use of pir sensor in the field does not show the difference whether predator in the field is human or animal So, it is limitation of this system